feat(aarch64): 定义独立的 guest CPU 状态
实现 AArch64 通用寄存器、向量寄存器、NZCV、浮点控制、TLS 与 exclusive monitor 状态模型。

公共 CPU 门面支持编译期选择架构，并通过 macOS、watchOS arm64_32 和 watchOS arm64 的状态语义测试。

+++ asbestos/asbestos.c
@@ -226,7 +226,8 @@ static int cpu_step_to_interrupt(struct cpu_state *cpu, struct tlb *tlb) {
         // block may be jetsam, but that's ok, because it can't be freed until
         // every thread on this asbestos is not executing anything
 
-        TRACE("%d %08x --- cycle %ld\n", current_pid(), ip, frame->cpu.cycle);
+        TRACE("%d %08x --- cycle %llu\n", current_pid(), ip,
+                (unsigned long long) frame->cpu.cycle);
 
         interrupt = fiber_enter(block, frame, tlb);
         if (interrupt == INT_NONE && __atomic_exchange_n(cpu->poked_ptr, false, __ATOMIC_SEQ_CST))
